requirements.txt
Flask==2.1.2
flask_bootstrap
requests
uuid
azure-cognitiveservices-speech
ffmpeg-python
Werkzeug==2.3.7
moviepy
pysrt
Flask-Reuploaded
auditok
transformers